Fix "codepoin_to_block" initial boundaries for binary search
"right = length" will lead to an error when searching for a value greater than
any of the ranges listed in the array. This error could not happen because the
last block "Supplementary Private Use Area-B" ends at U+10FFFF, and the function
"block_from_int" checks that the input is under 0x10FFFF. But it is better to
have a correct implementation of the binary search anyway.