Add "UnicodeData.txt" fetching and parsing
"UnicodeData.txt" is the definitive file from which most property
values are derived. The goal is to only use derived files (like
"DerivedName.txt") to test that there is no error in the derivation
algorithms. Using derivation rule to get the name of Hangul syllables
will remove almost a quarter of "name_map.mjs" (11 172 lines).